+++ packages/multichain-network-controller/src/index.ts
@@ -2,6 +2,9 @@ export { MultichainNetworkController } from './MultichainNetworkController/Multi
 export { MultichainNetworkService } from './MultichainNetworkService/MultichainNetworkService';
 export {
   getDefaultMultichainNetworkControllerState,
+  NON_EVM_TESTNET_IDS,
+  MULTICHAIN_NETWORK_TICKER,
+  MULTICHAIN_NETWORK_DECIMAL_PLACES,
   AVAILABLE_MULTICHAIN_NETWORK_CONFIGURATIONS,
 } from './constants';
 export type {