+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -32,6 +32,7 @@ import {
   syncInternalAccountsWithUserStorage,
 } from './account-syncing/controller-integration';
 import { setupAccountSyncingSubscriptions } from './account-syncing/setup-subscriptions';
+import { BACKUPANDSYNC_FEATURES } from './constants';
 import {
   performMainNetworkSync,
   startNetworkSyncing,
@@ -58,13 +59,19 @@ export type UserStorageControllerState = {
   /**
    * Condition used by UI and to determine if we can use some of the User Storage methods.
    */
-  isProfileSyncingEnabled: boolean | null;
+  isProfileSyncingEnabled: boolean;
   /**
    * Loading state for the profile syncing update
    */
   isProfileSyncingUpdateLoading: boolean;
   /**
-   * Condition used by E2E tests to determine if account syncing has been dispatched at least once.
+   * Condition used by UI to determine if account syncing is enabled.
+   */
+  isAccountSyncingEnabled: boolean;
+  /**
+   * Condition used to determine if account syncing has been dispatched at least once.
+   * This is used for event listeners to determine if they should be triggered.
+   * This is also used in E2E tests for verification purposes.
    */
   hasAccountSyncingSyncedAtLeastOnce: boolean;
   /**
@@ -84,6 +91,7 @@ export type UserStorageControllerState = {
 export const defaultState: UserStorageControllerState = {
   isProfileSyncingEnabled: true,
   isProfileSyncingUpdateLoading: false,
+  isAccountSyncingEnabled: true,
   hasAccountSyncingSyncedAtLeastOnce: false,
   isAccountSyncingReadyToBeDispatched: false,
   isAccountSyncingInProgress: false,
@@ -97,6 +105,10 @@ const metadata: StateMetadata<UserStorageControllerState> = {
   isProfileSyncingUpdateLoading: {
     persist: false,
     anonymous: false,
+  },
+  isAccountSyncingEnabled: {
+    persist: true,
+    anonymous: true,
   },
   hasAccountSyncingSyncedAtLeastOnce: {
     persist: true,
@@ -186,10 +198,6 @@ type ActionsObj = CreateActionsObj<
   | 'performDeleteStorage'
   | 'performBatchDeleteStorage'
   | 'getStorageKey'
-  | 'enableProfileSyncing'
-  | 'disableProfileSyncing'
-  | 'syncInternalAccountsWithUserStorage'
-  | 'saveInternalAccountToUserStorage'
 >;
 export type UserStorageControllerGetStateAction = ControllerGetStateAction<
   typeof controllerName,
@@ -211,14 +219,6 @@ export type UserStorageControllerPerformDeleteStorage =
 export type UserStorageControllerPerformBatchDeleteStorage =
   ActionsObj['performBatchDeleteStorage'];
 export type UserStorageControllerGetStorageKey = ActionsObj['getStorageKey'];
-export type UserStorageControllerEnableProfileSyncing =
-  ActionsObj['enableProfileSyncing'];
-export type UserStorageControllerDisableProfileSyncing =
-  ActionsObj['disableProfileSyncing'];
-export type UserStorageControllerSyncInternalAccountsWithUserStorage =
-  ActionsObj['syncInternalAccountsWithUserStorage'];
-export type UserStorageControllerSaveInternalAccountToUserStorage =
-  ActionsObj['saveInternalAccountToUserStorage'];
 
 export type AllowedActions =
   // Keyring Requests
@@ -283,7 +283,6 @@ export default class UserStorageController extends BaseController<
   // This is replaced with the actual value in the constructor
   // We will remove this once the feature will be released
   readonly #env = {
-    isAccountSyncingEnabled: false,
     isNetworkSyncingEnabled: false,
   };
 
@@ -342,7 +341,6 @@ export default class UserStorageController extends BaseController<
     state?: UserStorageControllerState;
     config?: ControllerConfig;
     env?: {
-      isAccountSyncingEnabled?: boolean;
       isNetworkSyncingEnabled?: boolean;
     };
     nativeScryptCrypto?: NativeScrypt;
@@ -354,7 +352,6 @@ export default class UserStorageController extends BaseController<
       state: { ...defaultState, ...state },
     });
 
-    this.#env.isAccountSyncingEnabled = Boolean(env?.isAccountSyncingEnabled);
     this.#env.isNetworkSyncingEnabled = Boolean(env?.isNetworkSyncingEnabled);
     this.#config = config;
 
@@ -391,15 +388,10 @@ export default class UserStorageController extends BaseController<
     this.#nativeScryptCrypto = nativeScryptCrypto;
 
     // Account Syncing
-    if (this.#env.isAccountSyncingEnabled) {
-      setupAccountSyncingSubscriptions(
-        { isAccountSyncingEnabled: true },
-        {
-          getUserStorageControllerInstance: () => this,
-          getMessenger: () => this.messagingSystem,
-        },
-      );
-    }
+    setupAccountSyncingSubscriptions({
+      getUserStorageControllerInstance: () => this,
+      getMessenger: () => this.messagingSystem,
+    });
 
     // Network Syncing
     if (this.#env.isNetworkSyncingEnabled) {
@@ -450,26 +442,6 @@ export default class UserStorageController extends BaseController<
     this.messagingSystem.registerActionHandler(
       'UserStorageController:getStorageKey',
       this.getStorageKey.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:enableProfileSyncing',
-      this.enableProfileSyncing.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:disableProfileSyncing',
-      this.disableProfileSyncing.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:syncInternalAccountsWithUserStorage',
-      this.syncInternalAccountsWithUserStorage.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:saveInternalAccountToUserStorage',
-      this.saveInternalAccountToUserStorage.bind(this),
     );
   }
 
@@ -635,43 +607,40 @@ export default class UserStorageController extends BaseController<
     return result;
   }
 
-  public async enableProfileSyncing(): Promise<void> {
+  public async setIsBackupAndSyncFeatureEnabled(
+    feature: keyof typeof BACKUPANDSYNC_FEATURES,
+    enabled: boolean,
+  ): Promise<void> {
     try {
       this.#setIsProfileSyncingUpdateLoading(true);
 
-      const isSignedIn = this.#auth.isSignedIn();
-      if (!isSignedIn) {
-        await this.#auth.signIn();
+      if (enabled) {
+        // If any of the features are enabled, we need to ensure the user is signed in
+        const isSignedIn = this.#auth.isSignedIn();
+        if (!isSignedIn) {
+          await this.#auth.signIn();
+        }
       }
 
       this.update((state) => {
-        state.isProfileSyncingEnabled = true;
-      });
+        if (feature === BACKUPANDSYNC_FEATURES.main) {
+          state.isProfileSyncingEnabled = enabled;
+        }
 
-      this.#setIsProfileSyncingUpdateLoading(false);
+        if (feature === BACKUPANDSYNC_FEATURES.accountSyncing) {
+          state.isAccountSyncingEnabled = enabled;
+        }
+      });
     } catch (e) {
-      this.#setIsProfileSyncingUpdateLoading(false);
       // istanbul ignore next
       const errorMessage = e instanceof Error ? e.message : JSON.stringify(e);
+      // istanbul ignore next
       throw new Error(
-        `${controllerName} - failed to enable profile syncing - ${errorMessage}`,
+        `${controllerName} - failed to ${enabled ? 'enable' : 'disable'} ${feature} - ${errorMessage}`,
       );
+    } finally {
+      this.#setIsProfileSyncingUpdateLoading(false);
     }
-  }
-
-  public async disableProfileSyncing(): Promise<void> {
-    const isAlreadyDisabled = !this.state.isProfileSyncingEnabled;
-    if (isAlreadyDisabled) {
-      return;
-    }
-
-    this.#setIsProfileSyncingUpdateLoading(true);
-
-    this.update((state) => {
-      state.isProfileSyncingEnabled = false;
-    });
-
-    this.#setIsProfileSyncingUpdateLoading(false);
   }
 
   #setIsProfileSyncingUpdateLoading(
@@ -718,7 +687,6 @@ export default class UserStorageController extends BaseController<
 
     await syncInternalAccountsWithUserStorage(
       {
-        isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled,
         maxNumberOfAccountsToAdd:
           this.#config?.accountSyncing?.maxNumberOfAccountsToAdd,
         onAccountAdded: () =>
@@ -747,14 +715,10 @@ export default class UserStorageController extends BaseController<
   async saveInternalAccountToUserStorage(
     internalAccount: InternalAccount,
   ): Promise<void> {
-    await saveInternalAccountToUserStorage(
-      internalAccount,
-      { isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled },
-      {
-        getMessenger: () => this.messagingSystem,
-        getUserStorageControllerInstance: () => this,
-      },
-    );
+    await saveInternalAccountToUserStorage(internalAccount, {
+      getMessenger: () => this.messagingSystem,
+      getUserStorageControllerInstance: () => this,
+    });
   }
 
   async syncNetworks() {