+++ packages/profile-sync-controller/src/controllers/user-storage/index.ts
@@ -5,5 +5,6 @@ export { Controller };
 export default UserStorageController;
 export * from './UserStorageController';
 export * as Mocks from './mocks';
+export * from './constants';
 export * from '../../shared/encryption';
 export * from '../../shared/storage-schema';