+++ packages/token-search-discovery-controller/src/token-search-discovery-controller.ts
@@ -14,6 +14,7 @@ import type {
   TrendingTokensParams,
   TopGainersParams,
   TopLosersParams,
+  SwappableTokenSearchParams,
 } from './types';
 
 // === GENERAL ===
@@ -145,6 +146,21 @@ export class TokenSearchDiscoveryController extends BaseController<
   ): Promise<TokenSearchResponseItem[]> {
     const results =
       await this.#tokenSearchService.searchTokens(tokenSearchParams);
+
+    this.update((state) => {
+      state.recentSearches = results;
+      state.lastSearchTimestamp = Date.now();
+    });
+
+    return results;
+  }
+
+  async searchSwappableTokens(
+    swappableTokenSearchParams: SwappableTokenSearchParams,
+  ): Promise<TokenSearchResponseItem[]> {
+    const results = await this.#tokenSearchService.searchSwappableTokens(
+      swappableTokenSearchParams,
+    );
 
     this.update((state) => {
       state.recentSearches = results;